refactor(components): rename style-helpers to style-mixins
