+++ src/components/BaseStyles/BaseStylesService.ts
@@ -16,7 +16,7 @@
 import { css } from '@emotion/core';
 
 import { StyleProps } from '../../styles/styled';
-import { textMega } from '../../styles/style-helpers';
+import { textMega } from '../../styles/style-mixins';
 
 const FONTS_BASE_URL = 'https://static.sumup.com/fonts/latin-greek-cyrillic';
