+++ src/components/LoadingButton/LoadingButton.tsx
@@ -18,7 +18,7 @@ import { FC } from 'react';
 import { css, jsx } from '@emotion/core';
 
 import styled, { StyleProps } from '../../styles/styled';
-import { hideVisually } from '../../styles/style-helpers';
+import { hideVisually } from '../../styles/style-mixins';
 import Button from '../Button';
 import { ButtonProps } from '../Button/Button';
 import Spinner from '../Spinner';
