+++ src/components/RadioButton/RadioButton.tsx
@@ -23,7 +23,7 @@ import {
   disableVisually,
   hideVisually,
   focusOutline,
-} from '../../styles/style-helpers';
+} from '../../styles/style-mixins';
 import { uniqueId } from '../../util/id';
 import useClickHandler from '../../hooks/use-click-handler';
