+++ src/components/Tabs/components/TabList/TabList.js
@@ -18,7 +18,7 @@ import PropTypes from 'prop-types';
 import styled from '@emotion/styled';
 import { css, jsx } from '@emotion/core';
 
-import { shadowDouble, hideScrollbar } from '../../../../styles/style-helpers';
+import { shadowDouble, hideScrollbar } from '../../../../styles/style-mixins';
 
 const MOBILE_AUTOSTRETCH_ITEMS_MAX = 3;
 const DEFAULT_HEIGHT = '80px';
