+++ src/components/Tag/Tag.tsx
@@ -20,7 +20,7 @@ import { Dispatch as TrackingProps } from '@sumup/collector';
 import { Theme } from '@sumup/design-tokens';
 
 import styled, { StyleProps } from '../../styles/styled';
-import { textMega, focusOutline } from '../../styles/style-helpers';
+import { textMega, focusOutline } from '../../styles/style-mixins';
 import useClickHandler from '../../hooks/use-click-handler';
 import { CloseButton, CloseButtonProps } from '../CloseButton/CloseButton';
